use uint80 for Registry.Instance extraData

+++ Fractal/contracts/modules/Registry.sol
@@ -136,7 +136,7 @@ contract Registry is Ownable {
 
     // instance state functions
 
-    function register(address instance, address creator, uint64 extraData) external {
+    function register(address instance, address creator, uint80 extraData) external {
         (
             FactoryStatus status,
             uint16 factoryID,
